move keyselectionvalueattribute to manager
