+++ src/es6/canvas_widget/KeySelectionValueListAttribute.js
@@ -5,7 +5,7 @@ import _ from "lodash-es";
 import IWCW from "../lib/IWCWrapper";
 
 import AbstractAttribute from "./AbstractAttribute";
-import KeySelectionValueAttribute from "./KeySelectionValueAttribute";
+import { KeySelectionValueAttribute } from "./Manager";
 import loadHTML from "../html.template.loader";
 import {
   AttributeAddOperation,
